+++ src/main/java/calculator/operations/CalculatorStrategy.java
@@ -2,6 +2,6 @@ package calculator.operations;
 
 public interface CalculatorStrategy {
 
-    int calculate(int[] numbers);
+    long calculate(long[] numbers);
 
 }
